+++ test/testall.py
@@ -50,6 +50,8 @@ def testall():
         ver_str = "%s.%s" % ver
         print "-- test with Python %s (%s)" % (ver_str, python)
         assert ' ' not in python
-        os.system("%s test.py -- -knownfailure" % python)
+        rv = os.system("%s test.py -- -knownfailure" % python)
+        if rv:
+            sys.exit(os.WEXITSTATUS(rv))
 
 testall()